Add : 작업현황 공유
- 고생하셨습니다.

+++ src/components/notice/CommentContainer.js
@@ -1,29 +1,43 @@
-import axios from "axios";
 import React, { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
+import axios from "axios";
+import ReactPaginate from "react-paginate";
+import { FiChevronLeft, FiChevronRight } from "react-icons/fi";
 import "./CommentContainer.scss";
+import "../../css/common/pagination.css";
+import PageNation from "../common/PageNation";
 
 const CommentContainer = () => {
   const { writerSeq } = useParams();
   const [comments, setComments] = useState([]);
+  const [currentPage, setCurrentPage] = useState(0); // 현재 페이지
+  const [totalPages, setTotalPages] = useState(0); // 전체 페이지 수
   const [formData, setFormData] = useState({
     boardSeq: writerSeq,
-    writer: 1,
+    writer: "",
     content: "",
   });
   const [editIndex, setEditIndex] = useState(-1); // 현재 수정 중인 댓글의 인덱스
   const [editContent, setEditContent] = useState(""); // 수정 중인 댓글 내용
 
   useEffect(() => {
-    fetchComments();
-  }, [writerSeq]);
+    const storedUser = JSON.parse(sessionStorage.getItem("user"));
+    const userSeq = storedUser?.userSeq;
+    if (!userSeq) {
+      alert("로그인 정보가 없습니다. 다시 로그인해주세요.");
+      return;
+    }
+    setFormData(prevFormData => ({ ...prevFormData, writer: userSeq }));
+    fetchComments(currentPage);
+  }, [writerSeq, currentPage]);
 
-  const fetchComments = async () => {
+  const fetchComments = async page => {
     try {
       const res = await axios.get(
-        `/api/community/comment?board_seq=${writerSeq}&page=1`,
+        `/api/community/comment?board_seq=${writerSeq}&page=${page + 1}`,
       );
       setComments(res.data.data.list);
+      setTotalPages(res.data.data.totalPages); // 전체 페이지 수 설정
     } catch (error) {
       console.error("Error fetching comments:", error);
     }
@@ -31,9 +45,17 @@ const CommentContainer = () => {
 
   const postComment = async () => {
     try {
+      if (!formData.writer) {
+        alert("로그인 정보가 없습니다. 다시 로그인해주세요.");
+        return;
+      }
+
       await axios.post("/api/community/comment", formData);
-      setFormData({ boardSeq: writerSeq, writer: 1, content: "" });
-      fetchComments();
+      setFormData(prevFormData => ({
+        ...prevFormData,
+        content: "",
+      }));
+      fetchComments(currentPage);
     } catch (error) {
       console.error("Error posting comment:", error);
     }
@@ -41,15 +63,22 @@ const CommentContainer = () => {
 
   const updateComment = async (commentSeq, content) => {
     try {
+      const storedUser = JSON.parse(sessionStorage.getItem("user"));
+      const userSeq = storedUser?.userSeq;
+      if (!userSeq) {
+        alert("로그인 정보가 없습니다. 다시 로그인해주세요.");
+        return;
+      }
+
       await axios.patch(`/api/community/comment`, null, {
         params: {
           commentSeq: commentSeq,
-          writer: 1,
+          writer: userSeq,
           content: content,
         },
       });
       setEditIndex(-1);
-      fetchComments();
+      fetchComments(currentPage);
     } catch (error) {
       console.error("Error updating comment:", error);
     }
@@ -57,13 +86,20 @@ const CommentContainer = () => {
 
   const deleteComment = async commentSeq => {
     try {
+      const storedUser = JSON.parse(sessionStorage.getItem("user"));
+      const userSeq = storedUser?.userSeq;
+      if (!userSeq) {
+        alert("로그인 정보가 없습니다. 다시 로그인해주세요.");
+        return;
+      }
+
       await axios.delete(`/api/community/comment`, {
         params: {
           commentSeq: commentSeq,
-          writer: 1,
+          writer: userSeq,
         },
       });
-      fetchComments();
+      fetchComments(currentPage);
     } catch (error) {
       console.error("Error deleting comment:", error);
     }
@@ -89,6 +125,12 @@ const CommentContainer = () => {
     event.preventDefault();
     updateComment(commentSeq, editContent);
   };
+
+  const handlePageChange = ({ selected }) => {
+    setCurrentPage(selected);
+  };
+
+  console.log(comments.length);
 
   return (
     <div>
@@ -140,6 +182,12 @@ const CommentContainer = () => {
           </div>
         </div>
       ))}
+
+      <PageNation
+        pageCount={2}
+        onPageChange={handlePageChange}
+        currentPage={currentPage}
+      />
 
       <form onSubmit={handleFormSubmit} className="comment__form">
         <textarea
